% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion}
\usage{
aic(loglik, k)
}
\arguments{
\item{loglik}{log-likelihood}

\item{k}{number of paramters}
}
\value{
AIC
}
\description{
Calculate the AIC for a model.
}
\examples{
gmm<-mixtools::normalmixEM(runif(1000))
aic(gmm$loglik,3)
}
