% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic_bt}
\alias{bic_bt}
\title{Bayesian Information Criterion (BIC)}
\usage{
bic_bt(loglik, k, n)
}
\arguments{
\item{loglik}{log-likelihood}

\item{k}{number of paramters}

\item{n}{sample size}
}
\value{
BIC
}
\description{
Calculate the BIC for a model.
}
\examples{
gmm<-mixtools::normalmixEM(runif(1000))
aic(gmm$loglik,3)
}
