% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btld-package.R
\docType{package}
\name{btld-package}
\alias{btld}
\alias{btld-package}
\title{btld: Triangular and Bimodal Triangular Linked Distributions}
\description{
This package contains functions to generate and calculate the pdf, cdf and icdfs of random variables following triangular and bimodal triangular-linked random variables. These functions are used to generate random variablesfrom the multivariate case using a Gaussian copula.
}
\author{
\strong{Maintainer}: Tristan Harris \email{th240498@gmail.com} (\href{https://orcid.org/0000-0002-4848-760X}{ORCID})

Authors:
\itemize{
  \item Alta de Waal \email{alta.dewaal@up.ac.za} (\href{https://orcid.org/0000-0001-8121-6249}{ORCID})
  \item Jocelyn Mazarura \email{jocelyn.mazarura@up.ac.za} (\href{https://orcid.org/0000-0003-4598-0834}{ORCID})
}

}
\keyword{internal}
