% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btld_scales.R
\name{btld_scales}
\alias{btld_scales}
\title{btld_scales}
\usage{
btld_scales(est_theta, F_thet)
}
\arguments{
\item{est_theta}{estimated breakpoint (mode) values from the H-plot}

\item{F_thet}{functional values at breakpoints}
}
\value{
estimated scale parameters
}
\description{
Function to determine the scale parameters from the H-plot
}
\examples{
th<-c(0.27,0.7)
pth <- c(0.605,0.85)
btld_scales(th,pth)
}
