% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btldParams.R
\name{btldParams}
\alias{btldParams}
\title{Determine the parameters of the btld}
\usage{
btldParams(xs, normalize = F)
}
\arguments{
\item{xs}{input data}

\item{normalize}{do we normalize?}
}
\value{
dataframe containing the scales and modes
}
\description{
Using the various functions built in this package, and a KDE to initialize the modal parameters, this function calculates the mode and scale parameters for the btld.
}
\examples{
library(mixtools)
data(NOdata)
btldParams(NOdata$NO,normalize=TRUE)
}
