% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btldParamsKDE.R
\name{btldParamsKDE}
\alias{btldParamsKDE}
\title{Determine the parameters of the btld}
\usage{
btldParamsKDE(xs, normalize = F)
}
\arguments{
\item{xs}{input data}

\item{normalize}{do we normalize?}
}
\value{
dataframe containing the scales and modes
}
\description{
Using the various functions built in this package, and a KDE to initialize the modal parameters, this function calculates the mode and scale parameters for the btld.
}
\examples{
library(mixtools)
data(NOdata)
btldParamsKDE(NOdata$NO, normalize=TRUE)
}
