% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtld.R
\name{dbtld}
\alias{dbtld}
\title{dbtld}
\usage{
dbtld(xvalues, alpha, theta)
}
\arguments{
\item{xvalues}{input values between 0 and 1}

\item{alpha}{scale vector parameter}

\item{theta}{Mode vector paramter}
}
\value{
df of pdf function values
}
\description{
Probability density function of the BTLD
}
\examples{

dbtld(runif(1000),  alpha = c(5,1), theta =c(0.3, 0.7))
}
