% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtridist.R
\name{dtridist}
\alias{dtridist}
\title{dtridist}
\usage{
dtridist(values, theta)
}
\arguments{
\item{values}{input vector}

\item{theta}{mode parameter (between 0 and 1)}
}
\value{
df of random variates
}
\description{
PDF for the triangular distribution
}
\examples{
dtridist(runif(1000),0.3)
dtridist(runif(100), 0.7)
}
