% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empcdf.R
\name{empcdf}
\alias{empcdf}
\title{empcdf}
\usage{
empcdf(df)
}
\arguments{
\item{df}{list of datapoints}
}
\value{
data.frame object with two columns: sorted values from input and ecdf
}
\description{
Function to determine the empirical cdf of data
}
\examples{
s<-seq(1,10,1)
empcdf(matrix(s))

}
