% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFThetas.R
\name{find.f_thetas}
\alias{find.f_thetas}
\title{Determine ecdf values for the mode parameters}
\usage{
find.f_thetas(x)
}
\arguments{
\item{x}{input data}
}
\value{
cdf values for the mode parameters
}
\description{
Search through the ecdf to find values that are around the modes and calculate the average density around the modes.
}
\examples{
x<-runif(1000,0,1)
find.f_thetas(x)
}
