% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findThetas.R
\name{find.thetas}
\alias{find.thetas}
\title{Determine modes of btld}
\usage{
find.thetas(x)
}
\arguments{
\item{x}{input data.}
}
\value{
modes
}
\description{
Using binned values, calculate the local maxima of the KDEs from the histograms. Then determine the midpoints of these densities.
}
\examples{
x<-runif(1000,5,10)
findThetas(x)
}
