% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findThetas.R
\name{findThetas}
\alias{findThetas}
\title{Determine modes of btld}
\usage{
findThetas(xs)
}
\arguments{
\item{xs}{input data.}
}
\value{
modes
}
\description{
Using kernel density estimation, calculate the local maxima i.e. the modes
}
\examples{
x<-runif(1000,0,1)
findThetas(x)
}
