% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_btld.R
\name{get_params_btld}
\alias{get_params_btld}
\title{get_params_btld}
\usage{
get_params_btld(data)
}
\arguments{
\item{data}{data that exhibits modes stored as a matrix}
}
\value{
parameters for a btld model
}
\description{
Function to determine the scale parameters from the H-plot for the multivariate case
}
\examples{
daa<- rbtld(size=1000, alpha = c(5,1), theta =c(0.3, 0.7))
get_params_btld(matrix(daa))

}
