% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_modeseeking.R
\name{knn_modeseeking}
\alias{knn_modeseeking}
\title{kNN Mode-Seeking Algorithm}
\usage{
knn_modeseeking(data, p = 0.15)
}
\arguments{
\item{data}{input data}

\item{p}{proportion of data to use as span}
}
\value{
matrix with 3 columns for the indices, data and modes.
}
\description{
This function calculates the knn modeseeking estimates which return the modes we use in the BTLD. The main reference for this function is:
}
\details{
Nordhaug Myhre, Jonas / Øyvind Mikalsen, Karl / Løkse, Sigurd / Jenssen, Robert
Robust clustering using a kNN mode seeking ensemble
2018
Pattern Recognition , Vol. 76
p. 491-505
}
\examples{
df<-rbtld(size=1000, alpha = c(5,1), theta =c(0.3, 0.7))
knn_modeseeking(df)

}
