% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn_cmvbtld.R
\name{mn_cmvbtld}
\alias{mn_cmvbtld}
\title{Random generation from the compound Multinomial-MVBTLD.}
\usage{
mn_cmvbtld(n, size, alpha, theta, sigma, dim)
}
\arguments{
\item{n}{size of dataset desired}

\item{size}{integer indicating the total number of objects to draw from in the multinomial}

\item{alpha}{scale matrix. Must be d-by-2 dimensions.}

\item{theta}{mode matrix. Must be d-by-2 dimensions.}

\item{sigma}{value on the off diagonal for Gaussian covariance matrix used in copula generation.}

\item{dim}{number of dimensions of copula. Must be same dimensions as the mode and scale parameters.}
}
\value{
matrix n by d+1 random variates from the compound MN-MVBTLD.
}
\description{
This function is a random generator for length n (d)-dimensional MVBTLD random variables. It also returns the density of these values.
}
\examples{
alpha<-matrix(c(3,3,3,3),nrow=2,ncol=2, byrow=TRUE)
theta<-matrix(c(0.3,0.7,0.3,0.7),ncol=2,nrow=2,byrow=TRUE)
mn_cmvbtld(100, size=10, alpha=alpha, theta=theta, sigma=0.5,dim=2)
}
