% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvbtld_scales.R
\name{mvbtld_scales}
\alias{mvbtld_scales}
\title{mvbtld_scales}
\usage{
mvbtld_scales(est_theta, F_thet)
}
\arguments{
\item{est_theta}{estimated breakpoint (mode) values stored as a vector for a d-dimensional model from the H-plot}

\item{F_thet}{functional values at breakpoints for each theta}
}
\value{
scale parameters for a mvbtld model
}
\description{
Function to determine the scale parameters from the H-plot for the multivariate case
}
\examples{
th<-matrix(c(0.3,0.7,0.3,0.7),ncol=2,nrow=2,byrow=TRUE)
pth <- matrix(c(0.17, 0.23, 0.43, 0.53),ncol=2,nrow=2,byrow=TRUE)
mvbtld_scales(est_theta=th,F_thet=pth)
}
