% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbtld.R
\name{pbtld}
\alias{pbtld}
\title{pbtld}
\usage{
pbtld(vec, alpha1, alpha2, theta1, theta2)
}
\arguments{
\item{vec}{input values}

\item{alpha1}{scale parameter 1}

\item{alpha2}{scale parameter 2}

\item{theta1}{Mode paramter 1.}

\item{theta2}{mode parameter 2.}
}
\value{
Vector of cumulative probabilities from the BTLD
}
\description{
Cumulative density function for the BTLD
}
\examples{
pbtld(runif(1000), alpha1 = 5, alpha2 = 1, theta1=0.3, theta2 = 0.7)
}
