% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbtld.R
\name{qbtld}
\alias{qbtld}
\title{quantile function for the btld}
\usage{
qbtld(x, theta, alpha)
}
\arguments{
\item{x}{A vector of inputs.}

\item{theta}{Mode vector}

\item{alpha}{Scale vector}
}
\value{
The icdf of \code{x}.
}
\description{
Quantile function (icdf) for the btld
}
\examples{
alpha<-c(5,1)
theta<-c(0.3,0.7)
qbtld(x=runif(1000), alpha, theta)
}
