% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtridist.R
\name{qtridist}
\alias{qtridist}
\title{qtridist}
\usage{
qtridist(values, theta)
}
\arguments{
\item{values}{input vector}

\item{theta}{mode parameter (between 0 and 1)}
}
\value{
df of random variates
}
\description{
icdf for the triangular distribution
}
\examples{
qtridist(runif(1000),0.3)
qtridist(runif(100), 0.7)
}
