% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbtld.R
\name{rbtld}
\alias{rbtld}
\title{Random number generator for the bimodal triangular linked distribution}
\usage{
rbtld(size, alpha, theta)
}
\arguments{
\item{size}{size of sample to generate.}

\item{alpha}{scale vector parameter}

\item{theta}{Mode vector paramter}
}
\value{
The random variates from a bltd
}
\description{
Using the icdf we generate random variates for the btld
}
\examples{
rbtld(size=1000, alpha = c(5,1), theta =c(0.3, 0.7))
}
