% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmvbtld.R
\name{rcmvbtld}
\alias{rcmvbtld}
\title{Random generation from the compositional MVBTLD.}
\usage{
rcmvbtld(n, theta, alpha, sigma, dim, ...)
}
\arguments{
\item{n}{size of dataset desired}

\item{theta}{mode matrix. 2 Must be d-by-2 dimensions.}

\item{alpha}{scale matrix. Must be d-by-2 dimensions.}

\item{sigma}{value on the off diagonal for Gaussian covariance matrix used in copula generation.}

\item{dim}{number of dimensions of copula. Must be same dimensions as the mode and scale parameters.}

\item{...}{kws for \code{normalCopula} and \code{rCopula}.}
}
\value{
matrix n by d+1 random variates from the compositional MVBTLD.
}
\description{
This function is a random generator for length n (d)-dimensional MVBTLD random variables. It is directly called in
the function \code{mu_mvbtld.R}.
}
\examples{
alpha<-matrix(c(3,3,3,3),nrow=2,ncol=2, byrow=TRUE)
theta<-matrix(c(0.3,0.7,0.3,0.7),ncol=2,nrow=2,byrow=TRUE)
rcmvbtld(1000, alpha=alpha, theta=theta, sigma=0.5,dim=2)
}
