% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvbtld.R
\name{rmvbtld}
\alias{rmvbtld}
\title{Random generation from MVBTLD}
\usage{
rmvbtld(n, theta, alpha, sigma, dim, normalize = FALSE, ...)
}
\arguments{
\item{n}{size of dataset desired}

\item{theta}{mode matrix. 2 Must be d-by-2 dimensions.}

\item{alpha}{scale matrix. Must be d-by-2 dimensions.}

\item{sigma}{value on the off diagonal for Gaussian covariance matrix used in copula generation. Can be vector.}

\item{dim}{number of dimensions}

\item{normalize}{default is FALSE. If true normalize across dimensions so that each observation is a probability space and thus sums to 1.}

\item{...}{kws for \code{normalCopula} and \code{rCopula}}
}
\value{
matrix n by d random variates from the MVBTLD
}
\description{
This function is the random generator for length n d-dimensional MVBTLD random variables.
}
\examples{
alpha<-matrix(c(3,3,3,3),nrow=2,ncol=2, byrow=TRUE)
theta<-matrix(c(0.3,0.7,0.3,0.7),ncol=2,nrow=2,byrow=TRUE)
rmvbtld(1000, alpha=alpha, theta=theta, sigma=0.5,dim=2)
}
