% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtridist.R
\name{rtridist}
\alias{rtridist}
\title{rtridist}
\usage{
rtridist(size, theta)
}
\arguments{
\item{size}{size of sample}

\item{theta}{mode parameter (between 0 and 1)}
}
\value{
df of random variates
}
\description{
random number generator for the triangular distribution
}
\examples{
rtridist(1000,0.3)
rtridist(100, 0.7)
}
