% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{softmax function}
\usage{
softmax(x, exp = T, base = NULL)
}
\arguments{
\item{x}{input data}

\item{exp}{logical to determine if to use exponential function}

\item{base}{base of softmax when \code{exp=F}}
}
\description{
This function computes the softmax normalized values for use in the multinomial
}
\examples{
alpha<-matrix(c(3,3,3,3),nrow=2,ncol=2, byrow=TRUE)
theta<-matrix(c(0.3,0.7,0.3,0.7),ncol=2,nrow=2,byrow=TRUE)
df<-rmvbtld(1000, alpha=alpha, theta=theta, sigma=0.5,dim=2)
softmax(df)
}
