% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridist_cdf.R
\name{tridist_cdf}
\alias{tridist_cdf}
\title{tridist_cdf}
\usage{
tridist_cdf(values, theta)
}
\arguments{
\item{theta}{mode parameter (between 0 and 1)}

\item{size}{size of sample}
}
\value{
df of random variates
}
\description{
random number generator for the triangular distribution
}
\examples{

tridist_cdf(runif(1000),0.3)
tridist_cdf(runif(1000), 0.7)
}
