% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitNormalize.R
\name{unit.normalize}
\alias{unit.normalize}
\title{Change scale of data to be over the range 0 to 1}
\usage{
unit.normalize(x)
}
\arguments{
\item{x}{input data}
}
\value{
data rescaled to lie between 0 and 1
}
\description{
Recsales data using the \code{caret} package
}
\examples{
x<-runif(1000,5,10)
unit.normalize(x)
}
