% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btldParams.R
\name{btld.params}
\alias{btld.params}
\title{Determine the parameters of the btld}
\usage{
btld.params(xs, normalize = F)
}
\arguments{
\item{xs}{input data}

\item{normalize}{do we normalize?}
}
\value{
dataframe containing the scales and modes
}
\description{
Using the various functions built in this package, this function calculates the mode and scale parameters for the btld.
}
\examples{
library(mixtools)
data(NOdata)
btld.params(NOdata$NO, normalize=TRUE)
}
