% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtld.R
\name{dbtld}
\alias{dbtld}
\title{dbtld}
\usage{
dbtld(xvalues, alpha1, alpha2, theta1, theta2)
}
\arguments{
\item{xvalues}{input values between 0 and 1}

\item{alpha1}{scale parameter}

\item{alpha2}{scale parameter}

\item{theta1}{scale parameter}

\item{theta2}{scale parameter}
}
\value{
df of pdf function values
}
\description{
Probability density function of the BTLD
}
\examples{

vals<-stats::runif(1000)
pdf <- dbtld(vals,  theta1 = 0.3, theta2 = 0.8, alpha1 = 3, alpha2 =3)
}
