% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findThetas.R
\name{find.thetas}
\alias{find.thetas}
\title{Determine modes of btld}
\usage{
find.thetas(xs)
}
\arguments{
\item{xs}{input data.}
}
\value{
modes
}
\description{
Using knn_modeseeking, calculate the local maxima.
}
\examples{
x<-runif(1000,0,1)
find.thetas(x)
}
