% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbtld.R
\name{pbtld}
\alias{pbtld}
\title{pbtld}
\usage{
pbtld(xvalues, alpha, theta)
}
\arguments{
\item{xvalues}{input values}

\item{alpha}{scale vector parameter 1}

\item{theta}{Mode vector paramter 1.}
}
\value{
Vector of cumulative probabilities from the BTLD
}
\description{
Cumulative distribution function for the BTL distribution
}
\examples{
pbtld(runif(1000), alpha = c(5,1), theta =c(0.3, 0.7))
}
