% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptridist.R
\name{ptridist}
\alias{ptridist}
\title{ptridist}
\usage{
ptridist(values, theta)
}
\arguments{
\item{values}{input vector}

\item{theta}{mode parameter (between 0 and 1)}
}
\value{
df of random variates
}
\description{
random number generator for the triangular distribution
}
\examples{

ptridist(runif(1000),0.3)
ptridist(runif(1000), 0.7)
}
