% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbtld.R
\name{qbtld}
\alias{qbtld}
\title{Quantile function (icdf) for the bltd}
\usage{
qbtld(x, theta1, theta2, alpha1, alpha2)
}
\arguments{
\item{x}{A vector of inputs.}

\item{theta1}{Mode paramter 1.}

\item{theta2}{mode parameter 2.}

\item{alpha1}{scale parameter 1}

\item{alpha2}{scale parameter 2}
}
\value{
The icdf of \code{x}.
}
\description{
Quantile function (icdf) for the bltd
}
\examples{
qbtld(x=runif(1000), alpha1 = 5, alpha2 = 1, theta1=0.3, theta2 = 0.7)
}
