% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbtld.R
\name{rbtld}
\alias{rbtld}
\title{rbtld}
\usage{
rbtld(size, alpha1, alpha2, theta1, theta2)
}
\arguments{
\item{size}{size of sample to generate.}

\item{alpha1}{scale parameter 1}

\item{alpha2}{scale parameter 2}

\item{theta1}{Mode paramter 1.}

\item{theta2}{mode parameter 2.}
}
\value{
The random variates from a bltd
}
\description{
Random number generator for the bltd
}
\examples{
rbtld(size=1000, alpha1 = 5, alpha2 = 1, theta1=0.3, theta2 = 0.7)
}
