% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbtld.R
\name{rbtld}
\alias{rbtld}
\title{rbtld}
\usage{
rbtld(size, alpha, theta)
}
\arguments{
\item{size}{size of sample to generate.}

\item{alpha}{scale vector parameter}

\item{theta}{Mode vector paramter}
}
\value{
The random variates from a bltd
}
\description{
Random number generator for the bltd
}
\examples{
rbtld(size=1000, alpha = c(5,1), theta =c(0.3, 0.7))
}
